Abstract more Edge code into the base UIA support, including a base UIA browseMode implementation, making it easier to offer Browse Mode for Win 10 Mail etc in future.
* Move some UIA utility functions off UIATextInfo and out of edge.py, into a new UIAUtils module.
* Add a UIABrowseMode module containing UIABrowseModeDocument which is a UIA-specific BrowseMode implementation. Edge is now based off of this. Edge.py now literally only contains Edge-Specific code.
